# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bug Tracker API",
        "description": "Dokumentasi REST API Bug Tracker untuk QA, DEV, dan PM.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/bugs": {
            "get": {
                "tags": [
                    "Bugs"
                ],
                "summary": "List bug (paginated)",
                "description": "Mengambil daftar bug dengan pagination dan optional filter status/severity.",
                "operationId": "39296d626dc6d8e1cd51bb483e29ca65",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status (OPEN, IN_PROGRESS, RESOLVED, CLOSED)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "OPEN"
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "description": "Filter severity (LOW, MEDIUM, HIGH, CRITICAL)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "HIGH"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar bug berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bugs"
                ],
                "summary": "Buat bug baru",
                "description": "QA melaporkan bug baru.",
                "operationId": "3ef01ac1accfc6842a8ebb40a2858c49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "reproduction_steps",
                                    "severity"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Login button tidak berfungsi"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tombol login tidak merespon."
                                    },
                                    "reproduction_steps": {
                                        "type": "string",
                                        "example": "1. Buka halaman login\\n2. Klik tombol login"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "HIGH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bug berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bugs/{id}": {
            "get": {
                "tags": [
                    "Bugs"
                ],
                "summary": "Detail bug",
                "description": "Mengambil detail satu bug lengkap dengan relasi.",
                "operationId": "d7b6039878a38708aa605ff58ef43d12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail bug berhasil diambil"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bugs"
                ],
                "summary": "Update data bug",
                "description": "Mengubah data bug (title, description, severity, assignee).",
                "operationId": "7e9c2362387eb4310d5aeac54d274eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Judul baru"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Deskripsi baru"
                                    },
                                    "reproduction_steps": {
                                        "type": "string",
                                        "example": "Step baru"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "example": "MEDIUM"
                                    },
                                    "assignee_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bug berhasil diupdate"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bugs"
                ],
                "summary": "Hapus bug",
                "description": "Menghapus bug dari sistem.",
                "operationId": "e3092df921f28a113c1ca342ccf74edb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bug berhasil dihapus"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bugs/{id}/status": {
            "put": {
                "tags": [
                    "Bugs"
                ],
                "summary": "Update status bug",
                "description": "Mengubah status bug dan mencatat history.",
                "operationId": "a6b0980bf23b316d7c8a3e2404102388",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bug",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "IN_PROGRESS"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Sedang dikerjakan oleh DEV"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diupdate"
                    },
                    "422": {
                        "description": "Status tidak valid / transition tidak diizinkan"
                    },
                    "404": {
                        "description": "Bug tidak ditemukan"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari /api/login atau /api/register.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}